requirements.txt
# ForexElite Pro - Python Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0

# Redis
redis==5.2.1
hiredis==3.0.0

# Trading & Data
oandapyV20==0.7.2
pandas==2.2.3
numpy==2.2.1
ta-lib==0.5.1

# FIX Protocol
asyncfix==1.0.1
simplefix==1.0.16

# Validation & Serialization
pydantic==2.10.5
pydantic-settings==2.7.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket
websockets==14.1

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1

# Code Quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.14.1

# Monitoring
prometheus-client==0.21.1
